+++ src/domain/flow-system/services/src/flow/flow_query_service_impl.rs
@@ -142,7 +142,7 @@ impl FlowQueryService for FlowQueryServiceImpl {
     ) -> Result<FlowStateListing, ListFlowsByDatasetError> {
         let total_count = self
             .flow_event_store
-            .get_count_flows_by_datasets(dataset_ids, &filters)
+            .get_count_flows_by_multiple_datasets(dataset_ids, &filters)
             .await?;
 
         let relevant_flow_ids: Vec<_> = self
@@ -165,27 +165,20 @@ impl FlowQueryService for FlowQueryServiceImpl {
     async fn list_all_datasets_with_flow_by_account(
         &self,
         account_id: &odf::AccountID,
-    ) -> Result<FlowDatasetListing, ListFlowsByDatasetError> {
+    ) -> Result<Vec<odf::DatasetID>, ListFlowsByDatasetError> {
+        // Consider using ReBAC: not just owned, but any relation to account
         let owned_dataset_ids = self
             .dataset_entry_service
             .get_owned_dataset_ids(account_id)
             .await
             .int_err()?;
 
-        let matched_stream = Box::pin(async_stream::try_stream! {
-            for dataset_id in &owned_dataset_ids {
-                let dataset_flows_count = self
-                    .flow_event_store
-                    .get_count_flows_by_dataset(dataset_id, &Default::default())
-                    .await?;
+        let owned_dataset_id_refs = owned_dataset_ids.iter().collect::<Vec<_>>();
 
-                if dataset_flows_count > 0 {
-                    yield dataset_id.clone();
-                }
-            }
-        });
-
-        Ok(FlowDatasetListing { matched_stream })
+        Ok(self
+            .flow_event_store
+            .filter_datasets_having_flows(&owned_dataset_id_refs)
+            .await?)
     }
 
     /// Returns states of system flows